refactor: move helper functions to separate file
